+++ lib/core/flame/components/entities/draggable_entity.dart
@@ -2,10 +2,12 @@ import 'dart:math';
 
 import 'package:defend_your_flame/constants/bounding_constants.dart';
 import 'package:defend_your_flame/constants/damage_constants.dart';
+import 'package:defend_your_flame/constants/misc_constants.dart';
 import 'package:defend_your_flame/constants/physics_constants.dart';
 import 'package:defend_your_flame/core/flame/components/entities/entity.dart';
 import 'package:defend_your_flame/core/flame/components/entities/entity_state.dart';
 import 'package:defend_your_flame/core/flame/managers/sprite_manager.dart';
+import 'package:defend_your_flame/helpers/misc_helper.dart';
 import 'package:defend_your_flame/helpers/physics_helper.dart';
 import 'package:defend_your_flame/helpers/timestep/timestep_helper.dart';
 import 'package:flame/components.dart';
@@ -124,7 +126,7 @@ class DraggableEntity extends Entity with DragCallbacks, GestureHitboxes {
 
     position += dragDistance;
 
-    position.y = position.y.clamp(BoundingConstants.minYCoordinate, _pickupHeight + 10);
+    position.y = position.y.clamp(BoundingConstants.minYCoordinate, _pickupHeight + MiscConstants.eps);
     position.x = position.x
         .clamp(BoundingConstants.minXCoordinateOffScreen, world.worldWidth + BoundingConstants.maxXCoordinateOffScreen);
