+++ lib/core/flame/components/entities/mobs/slime.dart
@@ -1,14 +1,13 @@
 import 'package:defend_your_flame/core/flame/components/entities/animation_config.dart';
+import 'package:defend_your_flame/core/flame/components/entities/disappear_on_death.dart';
 import 'package:defend_your_flame/core/flame/components/entities/draggable_entity.dart';
 import 'package:defend_your_flame/core/flame/components/entities/entity_config.dart';
 import 'package:defend_your_flame/core/flame/helpers/entity_helper.dart';
 import 'package:defend_your_flame/helpers/misc_helper.dart';
 import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
-import 'package:flame/effects.dart';
-import 'package:flutter/animation.dart';
 
-class Slime extends DraggableEntity {
+class Slime extends DraggableEntity with DisappearOnDeath {
   static final EntityConfig _slimeConfig = EntityConfig(
     entityResourceName: 'slime',
     defaultSize: Vector2(34, 27),
@@ -31,30 +30,10 @@ class Slime extends DraggableEntity {
     damageOnAttack: 3,
     goldOnKill: 4,
     walkingForwardSpeed: 42,
-    defaultScale: 1.05,
+    defaultScale: 1.15,
   );
 
-  bool _removingAnimation = false;
-
   Slime({super.scaleModifier}) : super(entityConfig: _slimeConfig);
-
-  @override
-  void update(double dt) {
-    if (!super.isAlive && !_removingAnimation) {
-      _removingAnimation = true;
-      add(OpacityEffect.by(
-        -0.95,
-        EffectController(
-          duration: 2,
-          curve: Curves.decelerate,
-        ),
-      )..onComplete = () {
-          isVisible = false;
-        });
-    }
-
-    super.update(dt);
-  }
 
   @override
   Vector2? attackEffectPosition() {
